+++ src/ru/strategy48/ejudge/standings/StandingsTableCell.java
@@ -9,12 +9,12 @@ public class StandingsTableCell {
     public boolean running = false;
 
     public int score = 0;
-    public int time = -1;
+    public long time = -1;
     public int attempts = 0;
 
     public boolean freezedSolved = false;
     public int freezedScore = 0;
-    public int freezedTime = -1;
+    public long freezedTime = -1;
     public int freezedAttempts = 0;
 
     public final Problem problem;
@@ -23,18 +23,18 @@ public class StandingsTableCell {
         this.problem = problem;
     }
 
-    public int getPenalty() {
+    public long getPenalty() {
         if (freezed) {
             if (!freezedSolved) {
                 return 0;
             } else {
-                return freezedTime / 60 + 20 * freezedAttempts;
+                return freezedTime / 60 + 20L * freezedAttempts;
             }
         } else {
             if (!solved) {
                 return 0;
             }
-            return time / 60 + 20 * attempts;
+            return time / 60 + 20L * attempts;
         }
     }
